+++ CHANGELOG.md
@@ -22,8 +22,10 @@
 - [BREAKING] Implement `LocalBlockProver` and rename `Block` to `ProvenBlock` (#1152, #1168, #1172).
 - [BREAKING] Added native types to `AccountComponentTemplate` (#1124).
 - Implemented `RemoteBlockProver`. `miden-proving-service` workers can prove blocks (#1169).
-- Use `Smt::with_entries` to error on duplicates in `StorageMap::with_entries` (#1167)
+- Used `Smt::with_entries` to error on duplicates in `StorageMap::with_entries` (#1167).
 - Implement user-facing bech32 encoding for `AccountId`s (#1185).
+- [BREAKING] Added `InitStorageData::from_toml()`, improved storage entry validations in `AccountComponentMetadata` (#1170).
+- [BREAKING] Rework miden-lib error codes into categories (#1196).
 - Added `info` log level by default in the proving service (#1200).
 - [BREAKING] Added port check in the proxy service to avoid starting up if the port is in use (#1200).